requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
httpx
feedparser
apscheduler
openai
google-generativeai
passlib[bcrypt]
python-jose[cryptography]
python-multipart